drm/nouveau: handle limited nvif ioctl in abi16
nouveau_usif.c was already stripped right back a couple of years ago,
limiting what userspace could do with it, and now I'd like to remove
the nvkm side of these interfaces entirely, in order to make it less
of a nightmare to add/change internal APIs in the future.

Unfortunately.  Userspace uses some of this.

Fortunately, userspace only ever ended up using a fraction of the APIs,
so I've reimplemented those in a more direct manner, and return -EINVAL
to userspace for everything else.

Signed-off-by: Ben Skeggs <bskeggs@nvidia.com>