nvkm/vgpu: set the VF partition count when NVIDIA vGPU is enabled
GSP firmware needs to know the number of max-supported vGPUs when
initialization.

The field of VF partition count in the GSP WPR2 is required to be set
according to the number of max-supported vGPUs.

Set the VF partition count in the GSP WPR2 when NVKM is loading the GSP
firmware and initializes the GSP WPR2, if vGPU is enabled.

Signed-off-by: Zhi Wang <zhiw@nvidia.com>