cxl: expose cxl_await_range_active()
Most of the CXL type-2 device don't have CXL memory device registers.

However, they can use CXL memory range registers in the configuration
space to tell the user if the device memory is ready or not.

Expose cxl_await_range_active() for the CXL type-2 device driver to
wait for the CXL memory is ready.

Signed-off-by: Zhi Wang <zhiw@nvidia.com>