cxl: introduce devm_cxl_del_memdev()
The teardown path of kernel CXL core heavily leverages the device
resource manager. Thus, the lifecycle of many created resources are
tied to the refcount of parent object and the resourced are freed
when the parent object is freed.

However, this creates a gap when an external caller wants to swept the
resource but keep the parent object for a re-initialization sequence.
E.g. in vfio-cxl.

Introduce the devm_cxl_del_memdev() for an external caller to destroy
the CXL memdev.

Signed-off-by: Zhi Wang <zhiw@nvidia.com>