drm/nova-core/vgpu: introduce FB memory allocation for vGPU
Creating a vGPU requires allocating a mgmt heap from the FB memory. The
size of the mgmt heap that a vGPU requires is from the vGPU type.

Expose the FB memory allocation to NVIDIA vGPU VFIO driver to allocate the
guest FB memory and the mgmt heap when creating a vGPU.

Signed-off-by: Zhi Wang <zhiw@nvidia.com>