drm/nova-core/vgpu: introduce BAR1 map routines for vGPUs
The mgmt heap is a block of shared FB memory between the GSP firmware
and the vGPU manager. It is used for supporting vGPU RPCs, vGPU logging.

To access the data structures of vGPU RPCs and vGPU logging, the mgmt
heap needs to mapped into BAR1 and the region in the BAR1 is required
to be mapped into CPU vaddr.

Expose the BAR1 map routines to NVIDIA vGPU VFIO driver to map the mgmt
heap.

Signed-off-by: Zhi Wang <zhiw@nvidia.com>