drm/nova-core/vgpu: introduce the PF event notification
The admin can unbind the PF driver or enable/disable VFs via the sysfs.

NVIDIA vGPU VFIO driver requires the PF driver to work. If the PF driver
is going to be unbound, the VFIO driver needs to be notified to teardown
first. Also, the VFIO driver needs to be notified to release the related
resources when admin disabling the VFs. To notify VFIO driver those
events, a PF event notification chain is introduced to broadcast events
to VFIO driver.

Introduce the PF event notification. Notify PCI SRIOV configuration and
PF driver unbind to the VFIO driver accordingly.

Signed-off-by: Zhi Wang <zhiw@nvidia.com>