drm/nova-core: calculate the checksum with updated GSP msg header
The page amount of a GSP command message is in the GSP msg header.

The code calculate the checksum before the page amount is finally
figured out. This is fine for a single-paged GSP command message as the
page amount is set to 1 by default. While the GSP command queue hangs
in the large GSP RPC when page amount > 1.

Update the calculated page amount to the GSP msg header before
calculating the checksum.

Signed-off-by: Zhi Wang <zhiw@nvidia.com>