drm/nova-core/vgpu: scrub the FB memory when vGPU support is enabled
To support the maximum vGPUs on the device that support vGPU, a larger
WPR2 heap size is required. On Ada with vGPU supported, the size should
be set to at least 581MB. By setting the WPR2 heap size up to 581MB, the
scrubber ucode image is required to scrub the FB memory before any other
ucode image is executed.

If not, the GSP firmware hangs when booting.

When vGPU support is enabled, reserve a large WPR2 heap size up to
581MB, execute the scrubber ucode image to scrub the FB memory before
executing any other ucode images, set the max supported VF to max in
WPR2 meta.

Signed-off-by: Zhi Wang <zhiw@nvidia.com>